Use strip-blobs-with-ids instead of delete-files
--delete-files removes files from old commits even in case the file was not modified.
This makes the latest commit look like as if it added all the png files.

The fix is to use strip-blobs-with-ids and pass only those blobs which are known to be unused now.

keep_blobs was built after --delete-files

Note: removal of "unused" pngs shrinks repository size from 80MB to 40MB